+++ tests/ecs-ec2/ecs-ec2.tf
@@ -12,11 +12,14 @@ provider "aws" {
 }
 
 module "ecs-ec2" {
-  source           = "../../modules/ecs-ec2"
-  ecs_cluster_name = var.ecs_cluster_name
-  image            = var.image
-  image_version    = var.image_version
-  memory           = var.memory
+  source                      = "../../modules/ecs-ec2"
+  ecs_cluster_name            = var.ecs_cluster_name
+  image                       = var.image
+  image_version               = var.image_version
+  image_mode                  = var.image_mode
+  supervised_image_repository = var.supervised_image_repository
+  supervised_image_version    = var.supervised_image_version
+  memory                      = var.memory
 
   coralogix_region = var.coralogix_region
   custom_domain    = var.custom_domain
@@ -25,12 +28,13 @@ module "ecs-ec2" {
   api_key_secret_arn = var.api_key_secret_arn
   api_key            = var.api_key
 
-  s3_config_bucket = var.s3_config_bucket
-  s3_config_key    = var.s3_config_key
+  s3_config_bucket         = var.s3_config_bucket
+  s3_config_key            = var.s3_config_key
+  s3_supervisor_config_key = var.s3_supervisor_config_key
 
   task_execution_role_arn = var.task_execution_role_arn
   task_role_arn           = var.task_role_arn
-  task_definition_arn      = var.task_definition_arn
+  task_definition_arn     = var.task_definition_arn
 
   health_check_enabled  = var.health_check_enabled
   health_check_interval = var.health_check_interval